habit-backend/requirements.txt
Django>=2.0,<3.0
psycopg2-binary==2.8
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
